requirements.txt
Django >= 1.4
django-registration
south
PIL
